Change amp to 0.
In py2, `amp` "leaks" out of the list comprehension and is used later.
In py3, this doesn't happen as list-comps have their own scope.
This was changed to 0 rather than the value of 16 that it probably would
have had before, as this is the only HDU that would contain BSCALE I think.
Moreover, I don't think this code did exactly what you wanted, and this should
be closer to the intended functionality, but I can't be sure, so you might want
to think about/check this.