Add new find() function that is more similar to old one.
This one stores things by name like it did before. I don't know which
you prefer, but this should be a drop-in replacement, just rename
findNew() to find(), delete the old one and you should be good to
go here (though I'd also give it a quick etst to make sure it behaves
the same).